Fix the user selected poll logic.
When the user has no selected item, the status of whether
or the user is hovering or an item is set to nil, and stored.
However, when fetching the status a nil is instead set to the
default value of false.  Since false ~= nil, this cause the
onTick loop to always update the gridlines.  If the user has
a potato laptop and a large blueprint selected, this will
significantly cut into UPS, and make multiplayer games
unplayable.

Also make the variables used in the check local, no need to
pollute the global namespace